1) Changed Test class name 2) Added test for partial download file delete

+++ downloaders/test_ftp_downloader.py
@@ -4,16 +4,18 @@ import unittest
 from downloaders import ftp_downloader
 from config.site_config import SiteConfig
 from unittest.mock import patch
+from unittest.mock import Mock
 
 
-class MyTest(TestCase):
+class TestFtpDownloader(TestCase):
 
     def setUp(self):
         self.fd = ftp_downloader.FtpDownloader(SiteConfig('protocol', 'url', 'host_name', 'user_name', 'password'), 'file')
 
     def tearDown(self):
         import os
-        os.remove("file")
+        if os.path.exists("file"):
+            os.remove("file")
 
     def test_download_file(self):
         with patch('downloaders.ftp_downloader.ftplib.FTP') as ftp_mock:
@@ -25,6 +27,18 @@ class MyTest(TestCase):
             mock_ftp.login.assert_called_with('user_name', 'password')
             self.assertTrue(mock_ftp.retrbinary.called)
             self.assertTrue(mock_ftp.quit.called)
+
+    def test_delete_partial_download_file(self):
+        with patch('downloaders.ftp_downloader.ftplib.FTP') as ftp_mock:
+            f = open('file', 'wb')
+            f.close()
+            loop = asyncio.get_event_loop()
+            mock_ftp = ftp_mock.return_value
+            mock_ftp.quit.side_effect = Mock(side_effect=Exception())
+            loop.run_until_complete(self.fd.download_and_save())
+            import os.path
+            self.assertFalse(os.path.exists('file'))
+
 
 
 if __name__ == '__main__':
